# microservices/high-five-service/src/api/api-spec.yaml
openapi: 3.0.1
info:
  title: high-five-service
  description: Service contract for high-five-service with OpenAPI 3.0 specifications
  version: 1.0.0
servers:
  - url: 'http://localhost:8000/high-fives/api'
  - url: 'http://localhost:8011/api'
  - url: 'http://18.217.249.183/high-fives/api'
...
security:
  - bearerAuth: []
  - x-userinfo: []
paths:
...
  /v1/highfive/{id}:
    get:
      tags:
        - High Fives
      summary: Get High Five by ID
      description: Returns a a single high five
      parameters:
        - name: id
          in: path
          description: ID of High Five Type
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HighFiveResponse'
        404:
          description: High Five not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - High Fives
      summary: Update a High Five
      description: updates a High Five
      parameters:
        - name: id
          in: path
          description: ID of High Five Type
          required: true
          schema:
            type: string
      requestBody:
        description: High Five Object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HighFiveBody'
        required: true
      responses:
        200:
          description: High Five Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HighFiveResponse'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: High Five not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-codegen-request-body-name: HighFiveBody
    delete:
      tags:
        - High Fives
      summary: Delete a High Five
      description: deletes a High Five
      parameters:
        - name: id
          in: path
          description: ID of High Five Type
          required: true
          schema:
            type: string
      responses:
        204:
          description: High Five was deleted
        404:
          description: High Five not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/highfive:
    get:
      tags:
        - High Fives
      summary: Get all High Fives
      description: Returns a list of High Fives
      parameters:
        - name: page
          in: query
          description: The number of the page
          schema:
            type: integer
        - name: pageSize
          in: query
          description: The numbers of items to return per page.
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HighFiveListResponse'
        default:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - High Fives
      summary: Create a High Five
      description: creates a High Five
      requestBody:
        description: High Five Details Object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HighFiveBody'
        required: true
      responses:
        201:
          description: High Five Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HighFiveResponse'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-codegen-request-body-name: HighFiveBody
...
  /v1/highfivetypes:
    get:
      tags:
        - High Five Types
      summary: Get all High Five Types
      description: Returns a list of High Five Types
      parameters:
        - name: page
          in: query
          description: The number of the page
          schema:
            type: integer
        - name: pageSize
          in: query
          description: The numbers of items to return per page.
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HighFiveTypeListResponse'
        default:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - High Five Types
      summary: Create a High Five Type
      description: Creates a High Five Type
      requestBody:
        description: highFiveType details
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HighFiveTypeBody'
        required: true
      responses:
        201:
          description: High Five Type Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HighFiveTypeResponse'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-codegen-request-body-name: HighFiveTypeBody
  /v1/highfivetypes/{id}:
    get:
      tags:
        - High Five Types
      summary: Get High Five Type by id
      description: Returns a high five type
      parameters:
        - name: id
          in: path
          description: ID of High Five Type
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HighFiveTypeResponse'
        404:
          description: High Five Type not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - High Five Types
      summary: Update a High Five Type
      description: Updates a High Five Type
      parameters:
        - name: id
          in: path
          description: ID of High Five Type
          required: true
          schema:
            type: string
      requestBody:
        description: Updated High Five Type
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HighFiveTypeBody'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HighFiveTypeResponse'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: High Five not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-codegen-request-body-name: HighFiveTypeBody
    delete:
      tags:
        - High Five Types
      summary: Delete a High Five Type
      description: Deletes a High Five Type
      parameters:
        - name: id
          in: path
          description: ID of High Five Type
          required: true
          schema:
            type: string
      responses:
        204:
          description: High Five Type was deleted
        404:
          description: High Five not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/highfivetypes/search:
    post:
      tags:
        - High Five Types
      summary: Search by High Five Types ID
      description: Searches for High Five Types given an array of ID's
      requestBody:
        description: High Five Type id's
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HighFiveTypeSearchBody'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HighFiveTypeListResponse'
        default:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-codegen-request-body-name: HighFiveTypeSearchBody
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: All requests against Kong Gateway must include the `Authorization` header containing your JWT Token.
    x-userinfo:
      type: apiKey
      in: header
      name: X-userinfo
      description: All requests against a standalone microservice must include the `x-userinfo` header containing your encoded user ID.
  schemas:
    ExtendedErrorInfo:
      required:
        - message
      type: object
      properties:
        message:
          type: string
        type:
          type: string
        key:
          type: string
        property:
          type: string
    Error:
      required:
        - message
      type: object
      properties:
        extraInfo:
          type: array
          items:
            $ref: '#/components/schemas/ExtendedErrorInfo'
        message:
          type: string
...
    HighFiveBody:
      required:
        - assignees
        - description
        - types
      properties:
        assignees:
          type: array
          items:
            type: string
        description:
          type: string
        notes:
          type: array
          items:
            type: object
            properties: {}
        types:
          type: array
          items:
            type: string
      example:
        assignees:
          - 7b3961e1-5397-467a-a81d-7766b9ea3093
          - 849cfb26-cba9-464f-8904-fc5c86b30aea
        description: some description
        types:
          - 849cfb26-cba9-464f-8904-fc5c86b30aea
          - 7b3961e1-5397-467a-a81d-7766b9ea3093
        attachments:
          images:
            - /s3/url1
            - /s3/url2
    HighFiveListResponse:
      required:
        - documents
        - metadata
      type: object
      properties:
        documents:
          type: array
          items:
            $ref: '#/components/schemas/HighFiveResponse'
        metadata:
          $ref: '#/components/schemas/MetaData'
    HighFiveResponse:
      required:
        - _createdBy
        - _createdOn
        - _id
        - _modifiedBy
        - _modifiedOn
        - assignees
        - description
        - notesCount
        - reactions
        - types
      properties:
        _id:
          type: string
        _createdOn:
          type: string
        _createdBy:
          type: string
        _modifiedOn:
          type: string
        _modifiedBy:
          type: string
        _deleted:
          type: boolean
        assignees:
          type: array
          items:
            type: string
        description:
          type: string
        reactions:
          type: object
          properties:
            currentUserReaction:
              type: string
              nullable: true
            star:
              type: integer
            clapping:
              type: integer
            party:
              type: integer
        attachments:
          type: object
          properties:
            images:
              type: array
              items:
                type: string
        notesCount:
          type: integer
        types:
          type: array
          items:
            type: string
      example:
        _id: a3b2d54f-e3d5-49a5-9fde-107ce4e34c91
        _createdOn: 2021-05-10T16:36:06.326Z
        _createdBy: 61087fbe-958a-4e21-865e-910c57141a91
        _modifiedOn: 2021-05-09T15:36:06.326Z
        _modifiedBy: c6681ac5-224c-479e-994b-44dbc3cc3f36
        _deleted: false
        assignees:
          - 7b3961e1-5397-467a-a81d-7766b9ea3093
          - 849cfb26-cba9-464f-8904-fc5c86b30aea
        description: some description
        types:
          - daefce2b-e05f-443f-932c-08166d9a8617
          - 7b3961e1-5397-467a-a81d-7766b9ea3093
        notesCount: 3,
        attachments:
          images:
            - /s3/url1
            - /s3/url2
        reactions:
          currentUserReaction: clapping
          clapping: 10
          party: 2
          star: 0
    HighFiveTypeBody:
      required:
        - name
      type: object
      properties:
        name:
          type: string
      example:
        name: Test Name for High Five Type
    HighFiveTypeSearchBody:
      required:
        - ids
      type: object
      properties:
        ids:
          type: array
          items:
            type: string
      example:
        ids:
          - a3b2d54f-e3d5-49a5-9fde-107ce4e34c91
          - 849cfb26-cba9-464f-8904-fc5c86b30aea
    HighFiveTypeListResponse:
      required:
        - documents
        - metadata
      type: object
      properties:
        documents:
          type: array
          items:
            $ref: '#/components/schemas/HighFiveTypeResponse'
        metadata:
          $ref: '#/components/schemas/MetaData'
    HighFiveTypeResponse:
      required:
        - _createdBy
        - _createdOn
        - _deleted
        - _id
        - _modifiedBy
        - _modifiedOn
        - name
      type: object
      properties:
        _id:
          type: string
        _createdOn:
          type: string
        _createdBy:
          type: string
        _modifiedOn:
          type: string
        _modifiedBy:
          type: string
        _deleted:
          type: boolean
        name:
          type: string
      example:
        _id: a3b2d54f-e3d5-49a5-9fde-107ce4e34c91
        _createdOn: 2021-05-10T16:36:06.326Z
        _createdBy: 61087fbe-958a-4e21-865e-910c57141a91
        _modifiedOn: 2021-05-09T15:36:06.326Z
        _modifiedBy: c6681ac5-224c-479e-994b-44dbc3cc3f36
        _deleted: false
        name: A High Five Type
...
    MetaData:
      required:
        - current_page_size
        - page_number
        - page_size
        - total_documents_count
        - total_pages
      type: object
      properties:
        total_documents_count:
          type: number
          description: Total number of documents available to read
        page_number:
          type: number
          description: Current Page Number
        page_size:
          type: number
          description: Maximum number of elements per page
        current_page_size:
          type: number
          description: Total elements on the current page requested
        total_pages:
          type: number
          description: Total